Add: Interface to set up resummation in usercode
Moved the set up of the resummation into usercode and returned
chooser.f to the default. In reader check to make sure the
resummation can be performed on the process in question rather
than allowing it to be performed on all processes. Change the
kprocess.f file to be consistent with fortran 77 and 90.