Refactor: Move JetVHeto components into a library
Moved all of the Sudakov calculations into a static library. MCFM
compiles as it should and retains all previous functionality but is
less tightly coupled. Removed some remnants of the old method of
calculation in resmset. Decoupled some sections of the code further,
dipoles still needs some work but other parts of the code are good.