Fix: Corrected bug in how born_config is set
When born_config was moved into jetvheto.f it was added to the
jetvhetovars which is given the omp directive threadprivate. This
made it limited to just the first thread as far as I can tell. I
think the setup is done on a single thread so then only this thread
has access and is able to set up other parameters needed for
virtfin to be run correctly.

Also corrected makefile.orig.