Add: Switch to compute signal interference
Allow the user to ask for just the interference between Higgs
production terms i.e. top loop and ggh contact interaction. This
could otherwise not be set independently just by changing the
values of the kappa coupling rescalings. This is added to the input
card with a logical parameter, so it can be selected.